" MetricsController"

+++ BankApp/banking_app/src/main/java/revature/banking_app/Data/iDatabase.java
@@ -42,7 +42,7 @@ public interface iDatabase {
       String canceledAccount = "canceled";
 
       String accountBalance = "balance";
-      int defaultBalance = 0;
+      long defaultBalance = 0;
 
      String owners = "owners";
